# api-service/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "api-service",
		"description": "guggr api definition",
		"contact": {
			"name": "Guggr"
		},
		"license": {
			"name": ""
		},
		"version": "0.1.0"
	},
	"paths": {
		"/api/v1/auth/login": {
			"post": {
				"tags": ["auth"],
				"summary": "login endpoint",
				"operationId": "auth_login",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Access- and refresh-token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authorized"
					},
					"500": {
						"description": "Storage error"
					}
				}
			}
		},
		"/api/v1/auth/logout": {
			"post": {
				"tags": ["auth"],
				"summary": "logout endpoint",
				"operationId": "auth_logout",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TokenRefreshRequest"
							}
						}
					},
					"required": true
				},
				"responses": {
					"204": {
						"description": "Successful logout"
					},
					"500": {
						"description": "Storage error"
					}
				}
			}
		},
		"/api/v1/auth/token/refresh": {
			"post": {
				"tags": ["auth"],
				"summary": "token refresh endpoint",
				"operationId": "auth_refresh_token",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TokenRefreshRequest"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Access- and refresh-token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authorized"
					},
					"500": {
						"description": "Storage error"
					}
				}
			}
		},
		"/api/v1/groups": {
			"get": {
				"tags": ["groups"],
				"summary": "list endpoint for groups",
				"operationId": "list_group",
				"responses": {
					"200": {
						"description": "List groups",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/DisplayGroup"
									}
								}
							}
						}
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"post": {
				"tags": ["groups"],
				"summary": "create endpoint for groups",
				"operationId": "create_group",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateGroup"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Created group",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DisplayGroup"
								}
							}
						}
					},
					"400": {
						"description": "Validation error"
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
		"/api/v1/groups/{id}": {
			"get": {
				"tags": ["groups"],
				"summary": "get endpoint for groups",
				"operationId": "get_group",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Group id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Group",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DisplayGroup"
								}
							}
						}
					},
					"404": {
						"description": "Group Not Found"
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"delete": {
				"tags": ["groups"],
				"summary": "delete endpoint for groups",
				"operationId": "delete_group",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Group id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Deleted"
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"patch": {
				"tags": ["groups"],
				"summary": "update endpoint for groups",
				"operationId": "update_group",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Group id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateGroup"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Patched group",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DisplayGroup"
								}
							}
						}
					},
					"400": {
						"description": "Validation error"
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
			"CreateGroup": {
				"type": "object",
				"description": "sent to the database for creating a new group",
				"required": ["name"],
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1
					}
				}
			},
...
			"DisplayGroup": {
				"type": "object",
				"description": "returned from the database when displaying a group",
				"required": ["id", "name"],
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				}
			},
...
			"LoginRequest": {
				"type": "object",
				"description": "request by the user to login",
				"required": ["email", "password"],
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				}
			},
			"TokenRefreshRequest": {
				"type": "object",
				"description": "request by the user to get a new [`TokenResponse`]",
				"required": ["refresh_token"],
				"properties": {
					"refresh_token": {
						"type": "string"
					}
				}
			},
			"TokenResponse": {
				"type": "object",
				"description": "returned to the user on login / token refresh",
				"required": ["access_token", "refresh_token"],
				"properties": {
					"access_token": {
						"type": "string"
					},
					"refresh_token": {
						"type": "string"
					}
				}
			},
			"UpdateGroup": {
				"type": "object",
				"description": "sent to the database when updating a group",
				"properties": {
					"name": {
						"type": ["string", "null"],
						"minLength": 1
					}
				}
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}